Fix carousel index two-way binding and delimiters colour.
