+++ src/MudBlazor/Base/MudBaseInput.cs
@@ -373,7 +373,7 @@ namespace MudBlazor
                     // Text update suppression, only in BSS (not in WASM).
                     // This is a fix for #1012
                     if (RuntimeLocation.IsServerSide && TextUpdateSuppression)
-                        updateText=false;
+                        updateText = false;
                 }
                 if (updateText)
                 {
@@ -406,7 +406,7 @@ namespace MudBlazor
 
         protected override void ResetValue()
         {
-            SetTextAsync(null, updateValue:true).AndForget();
+            SetTextAsync(null, updateValue: true).AndForget();
             base.ResetValue();
         }
     }
