+++ src/MudBlazor/Base/MudBaseItemsControl.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 
@@ -35,6 +34,7 @@ namespace MudBlazor
 
                 LastContainer = _selectedIndexField >= 0 ? SelectedContainer : null;
                 _selectedIndexField = value;
+                SelectionChanged();
                 StateHasChanged();
                 SelectedIndexChanged.InvokeAsync(value);
             }
@@ -106,31 +106,9 @@ namespace MudBlazor
                 SelectedIndex = index;
             }
         }
-    }
 
-    public abstract class MudBaseBindableItemsControl<TChildComponent, TData> : MudBaseItemsControl<TChildComponent>
-        where TChildComponent : MudComponentBase
-    {
+        protected virtual void SelectionChanged() { }
 
-        /// <summary>
-        /// Items Collection - For databinding usage
-        /// </summary>
-        [Parameter]
-        public IEnumerable<TData> ItemsSource { get; set; }
-
-        /// <summary>
-        /// Template for each Item in ItemsSource collection
-        /// </summary>
-        [Parameter]
-        public RenderFragment<TData> ItemTemplate { get; set; }
-
-        /// <summary>
-        /// Gets the Selected Item from ItemsSource, or Selected TChildComponent, when it's null
-        /// </summary>
-        public object SelectedItem
-        {
-            get => ItemsSource == null ? Items[SelectedIndex] : ItemsSource.ElementAtOrDefault(SelectedIndex);
-        }
-
+        public virtual void AddItem(TChildComponent item) { }
     }
 }
