+++ src/MudBlazor/Components/Carousel/MudCarouselItem.razor.cs
@@ -69,7 +69,7 @@ namespace MudBlazor
 
         protected override Task OnInitializedAsync()
         {
-            Parent?.Items.Add(this);
+            Parent?.AddItem(this);
             return Task.CompletedTask;
         }
